requirements.txt
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.6.0
MarkupSafe==0.23
pycryptodome==3.4.11
pymongo==3.4.0
requests==2.13.0
Werkzeug==0.11.15
